requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
duckdb>=1.1.3
google-genai>=0.2.0
pydantic>=2.6.0
python-dotenv>=1.0.1
feedparser>=6.0.11
beautifulsoup4>=4.12.3
requests>=2.31.0